+++ src/index.ts
@@ -14,7 +14,7 @@ import * as PerformanceService from './service/performance';
 import * as ReserveService from './service/reserve';
 import * as StockService from './service/stock';
 import * as TaskService from './service/task';
-import * as PlaceOrderTransactionService from './service/transaction/placeOrder';
+// import * as PlaceOrderTransactionService from './service/transaction/placeOrder';
 import * as PlaceOrderInProgressTransactionService from './service/transaction/placeOrderInProgress';
 import * as ReturnOrderTransactionService from './service/transaction/returnOrder';
 
@@ -56,7 +56,7 @@ export namespace service {
     export import stock = StockService;
     export import task = TaskService;
     export namespace transaction {
-        export import placeOrder = PlaceOrderTransactionService;
+        export import placeOrder = cinerino.service.transaction.placeOrder;
         export import placeOrderInProgress = PlaceOrderInProgressTransactionService;
         export import returnOrder = ReturnOrderTransactionService;
     }