+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 座席予約承認アクションのobjectに進行中予約取引を追加
 - 取引リポジトリをCinerino化
 - Chevre@6.0.0対応
+- 注文取引サービスを@cinerino/domainへ移行
 
 ### Deprecated